refactor: remove duplicate helper function
