+++ src/services/blog/blog.js
@@ -1,6 +1,7 @@
 import redisClient from '../../utils/redis.js';
 import { getTheDevRead } from './request.js';
-import { randomArray, renderTemplate } from './utils.js';
+import { randomArray } from './utils.js';
+import { renderTemplate } from '../../utils/template.js';
 import { getCommandArgs } from '../../utils/command.js';
 
 const WHITELIST = ['javascript', 'php', 'go', 'c', 'typescript', 'python'];
@@ -30,7 +31,9 @@ async function devRead(context, cache) {
 
     if (queryData) {
       const items = randomArray(JSON.parse(queryData), 3);
-      const read = items.map(({ title, body, url }) => renderTemplate({ title, body, url })).join('\n');
+      const read = items
+        .map(({ title, body, url }) => renderTemplate('blog/template.hbs', { title, body, url }))
+        .join('\n');
       await context.telegram.sendMessage(context.message.chat.id, read, { parse_mode: 'HTML' });
       return;
     }
@@ -46,7 +49,7 @@ async function devRead(context, cache) {
 
   const items = randomArray(data, 3);
   const read = items
-    .map((x) => renderTemplate({ title: x?.title ?? '', body: x?.body ?? '', url: x?.url ?? '' }))
+    .map((x) => renderTemplate('blog/template.hbs', { title: x?.title ?? '', body: x?.body ?? '', url: x?.url ?? '' }))
     .join('\n');
 
   await context.telegram.sendMessage(context.message.chat.id, read, { parse_mode: 'HTML' });
